requirements.txt
amadeus==11.0.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudinary==1.41.0
cryptography==44.0.0
defusedxml==0.7.1
diff-match-patch==20241021
dj-rest-auth==7.0.1
Django==5.1.3
django-allauth==65.3.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-import-export==4.2.0
django-jazzmin==3.0.1
django-unfold==0.40.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==24.1
paypalrestsdk==1.13.3
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.1
tablib==3.7.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
